requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
pydantic>=2.5.0
pymongo>=4.6.0
redis>=5.0.1
celery>=5.3.4
stem>=1.8.2
scapy>=2.5.0
networkx>=3.2.1
pandas>=2.1.4
numpy>=1.26.0
scikit-learn>=1.3.2
matplotlib>=3.8.2
plotly>=5.17.0
requests>=2.31.0
aiohttp>=3.9.1
langchain>=0.1.0
langchain-google-genai>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-socketio>=5.10.0
dnspython>=2.4.2
maxminddb>=2.2.0
geoip2>=4.7.0
python-dotenv>=1.0.0
motor>=3.3.0
pydantic-settings>=2.0.0